Made controlling of seperate units Possible
Reorganized scripts so there is a mouseManager that checks to see if you clicked with the mouse and what code should be activated, then activate that code. To do this I added hitObject to see the name of the object hit (or if something at all was hit) and used that to check that object for the file with the code I want to start.