Added Weapons, Arrow Code, and Identification of enemy click
I added simple models for weapons and added them to the prefabs of the units. I also added a cube with the tag enemy. This will have to changed later on to make multiple players. I also added an arrow prefab that has code that is not called. We first need an Identifier for when to attack. Lastely made a Debug.log statement for when you have a unit selected and you right click an enemy.