Added Screen Movement, Control Groups, and Double Click
Made the Screen move if you put you mouse to the side of the screen. Currently not polished. Also have control groups now. Just needs to be optimized in script. Lastely added so you can double click a unit and it will select all units of the same time on the screen.